Added a patch by Adam Pribyl (the original author of wwsr) to support weatherstats type output as well.
Adams own words about the patch:
"As the wwsr I was using was really a hack and I consider wsp much better in all aspects, I patched the wsp to output the data in weatherstats format as weatherstats depends on the text names for values from wwsr/wsp to create all the outputs.

Attached is a patch to the existing SVN trunk of wsp. It adds option --weatherstats that outputs the data in weatherstats formating. It also needs -c 2 to calculate the rain 1h. I know there were some complains about this in wwsr, but I am not sure what is wrong with such a calculation. It only does
((last stored value - current read value) * 60/delay from last stored value)
to calculate approximate rain withing 1h. I know it is not perfect, but it is at least some measure of how heavy the rain is."